roles de usuario

+++ creditos3a/app/controllers/admin/categorias_controller.rb
@@ -1,5 +1,6 @@
 class Admin::CategoriasController < Admin::AdminController
   before_action :set_categoria, only: [:show, :edit, :update, :destroy]
+  before_action :check_admin
 
   def index
     @categorias = Categoria.all
@@ -18,6 +19,8 @@ class Admin::CategoriasController < Admin::AdminController
 
   def create
     @categoria = Categoria.new(categoria_params)
+    authorize :admin_categories?
+
     if @categoria.save
       flash[:success] = 'Categoría creada exitosamente.'
       redirect_to admin_categorias_path
@@ -27,6 +30,8 @@ class Admin::CategoriasController < Admin::AdminController
   end
 
   def update
+    authorize :admin_categories?
+
     if @categoria.update(categoria_params)
       redirect_to admin_categoria_path(@categoria)
     else
@@ -35,6 +40,8 @@ class Admin::CategoriasController < Admin::AdminController
   end
 
   def destroy
+    authorize :admin_categories?
+
     if @categoria.destroy
       redirect_to admin_categorias_path
     else
@@ -50,5 +57,11 @@ class Admin::CategoriasController < Admin::AdminController
 
   def categoria_params
     params.require(:categoria).permit(:id, :nombre, :imagen, :producto_id, :catalogo_id, :producto_nombre, :producto_descripcion, :producto_file, :producto_avatar)
+  end
+
+  def check_admin
+    unless current_user.administrador?
+      redirect_to root_path, alert: 'No tienes permisos para acceder a esta página.'
+    end
   end
 end
